+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.0] - 2017-08-25
+- Added static method `hash`
+- Deprecated static methods `MD5` and `SHA1` 
+
 ## [1.2.0] - 2017-06-10
 ### Added
-- Static methods MD5 and SHA1 
+- Static methods `MD5` and `SHA1` 
 
 ## [1.1.0] - 2017-06-07
 ### Added